Fixed entities and repositories

+++ src/main/java/com/example/sap_shop/repository/CategoryRepository.java
@@ -3,5 +3,6 @@ package com.example.sap_shop.repository;
 import com.example.sap_shop.model.Category;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CategoryRepository extends CrudRepository<Category, Integer> {
+public interface CategoryRepository extends CrudRepository<Category, Long> {
+
 }
