+++ src/main/java/com/example/sap_shop/repository/OrderItemRepository.java
@@ -3,5 +3,5 @@ package com.example.sap_shop.repository;
 import com.example.sap_shop.model.OrderItem;
 import org.springframework.data.repository.CrudRepository;
 
-public interface OrderItemRepository extends CrudRepository<OrderItem, Integer> {
+public interface OrderItemRepository extends CrudRepository<OrderItem, Long> {
 }
